Fixed missing cast in test case

+++ src/unitTests/ttCreation.cxx
@@ -104,9 +104,9 @@ static misc::UnitTest tt_crea_eps("TT", "creation_with_epsilon", [](){
 	size_t numDecrease = 5-ttA.rank(0) + 25 - ttA.rank(1) + 5 - ttA.rank(2);
 	Index i;
 	
-	TEST(frob_norm(A(i&0)-Tensor(ttA)(i&0))/frob_norm(A) < numDecrease*EPS);
+	TEST(frob_norm(A(i&0)-Tensor(ttA)(i&0))/frob_norm(A) < static_cast<double>(numDecrease)*EPS);
 	TEST(ttA.ranks()[1] < 25);
-	TEST(frob_norm(A(i&0)-Tensor(ttB)(i&0))/frob_norm(A) < numDecrease*EPS);
+	TEST(frob_norm(A(i&0)-Tensor(ttB)(i&0))/frob_norm(A) < static_cast<double>(numDecrease)*EPS);
 	TEST(ttB.ranks()[1] < 25);
 });
